// types/categoires.d.ts
// Generated by https://quicktype.io

export interface Categories {
  categories: Category[];
}

export interface Category {
  __typename: Typename;
  slug: string;
  name: string;
  arabicDisplay: string;
  subCategories?: Category[];
}

export enum Typename {
  Category = "Category",
  SubCategory = "SubCategory",
}